requirements.txt
flask_cors
dotenv
qrcode
pillow
numpy
rapidocr_onnxruntime
ultralytics
openpyxl
